requirements.txt
aniso8601==9.0.1
click==8.1.2
Flask==2.1.1
Flask-RESTful==0.3.9
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
typing_extensions==4.2.0
Werkzeug==2.1.1
zipp==3.8.0
